Add a big commit of moving file structure around
